Fix deferred-save lint checks
Keep generated test identifiers lossless and split lifecycle tests so the
focused coverage remains within the repository module-size policy.
